requirements.txt
boto3==1.4.7
botocore==1.7.1
click==6.7
docutils==0.14
Flask==0.12.2
Flask-SQLAlchemy==2.2
Flask-SSLify==0.1.5
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
MarkupSafe==1.0
python-dateutil==2.6.1
python-dotenv==0.6.5
s3transfer==0.1.10
six==1.10.0
SQLAlchemy==1.1.13
Werkzeug==0.12.2
